requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
python-dotenv==1.0.1
gunicorn==21.2.0
aiosqlite==0.20.0
